Move private parsers out of main module
