+++ src/parser/component.rs
@@ -2,18 +2,44 @@ use std::iter::once;
 
 use winnow::{
     ascii::{newline, space0},
-    combinator::{delimited, opt, preceded, repeat, separated_pair, terminated},
+    combinator::{alt, delimited, opt, preceded, repeat, separated_pair, terminated},
     error::ParserError,
     token::{one_of, take_while},
-    Parser,
+    PResult, Parser,
 };
 
-use crate::{Attribute, Name, Value};
+use crate::{Attribute, Name, Object, Value};
+
+/// Parse an object with at least one attribute terminated by a newline.
+///
+/// As per [RFC 2622](https://datatracker.ietf.org/doc/html/rfc2622#section-2), an RPSL object
+/// is textually represented as a list of attribute-value pairs that ends when a blank line is encountered.
+pub fn object_block<'s>(input: &mut &'s str) -> PResult<Object<'s>> {
+    let (attributes, source) = terminated(repeat(1.., attribute()), newline)
+        .with_taken()
+        .parse_next(input)?;
+    Ok(Object::from_parsed(source, attributes))
+}
+
+/// Extends the object block parser to consume optional padding server messages or newlines.
+pub fn object_block_padded<'s>(input: &mut &'s str) -> PResult<Object<'s>> {
+    delimited(
+        consume_opt_message_or_newlines,
+        object_block,
+        consume_opt_message_or_newlines,
+    )
+    .parse_next(input)
+}
+
+/// Consume optional server messages or newlines.
+fn consume_opt_message_or_newlines(input: &mut &str) -> PResult<()> {
+    repeat(0.., alt((newline.void(), server_message().void()))).parse_next(input)
+}
 
 // A response code or message sent by the whois server.
 // Starts with the "%" character and extends until the end of the line.
 // In contrast to RPSL, characters are not limited to ASCII.
-pub fn server_message<'s, E>() -> impl Parser<&'s str, &'s str, E>
+fn server_message<'s, E>() -> impl Parser<&'s str, &'s str, E>
 where
     E: ParserError<&'s str>,
 {
@@ -26,7 +52,7 @@ where
 
 // Generate an attribute parser.
 // The attributes name and value are separated by a colon and optional spaces.
-pub fn attribute<'s, E>() -> impl Parser<&'s str, Attribute<'s>, E>
+fn attribute<'s, E>() -> impl Parser<&'s str, Attribute<'s>, E>
 where
     E: ParserError<&'s str>,
 {
@@ -105,6 +131,74 @@ mod tests {
     use winnow::error::ContextError;
 
     use super::*;
+
+    #[test]
+    fn object_block_valid() {
+        let object = &mut concat!(
+            "email:       rpsl-rs@github.com\n",
+            "nic-hdl:     RPSL1-RIPE\n",
+            "\n"
+        );
+        assert_eq!(
+            object_block(object),
+            Ok(Object::from_parsed(
+                object,
+                vec![
+                    Attribute::unchecked_single("email", "rpsl-rs@github.com"),
+                    Attribute::unchecked_single("nic-hdl", "RPSL1-RIPE")
+                ]
+            ))
+        );
+    }
+
+    #[test]
+    /// When parsing RPSL, the resulting object contains the original source it was created from.
+    fn parsed_object_contains_source() {
+        let rpsl = &mut concat!(
+            "email:       rpsl-rs@github.com\n",
+            "nic-hdl:     RPSL1-RIPE\n",
+            "\n"
+        );
+        let source = *rpsl;
+        let object = object_block(rpsl).unwrap();
+        assert_eq!(object.source().unwrap(), source);
+    }
+
+    #[test]
+    fn object_block_without_newline_termination_is_err() {
+        let object = &mut concat!(
+            "email:       rpsl-rs@github.com\n",
+            "nic-hdl:     RPSL1-RIPE\n",
+        );
+        assert!(object_block(object).is_err());
+    }
+
+    #[rstest]
+    #[case(
+        &mut "% Note: This is a server message\n"
+    )]
+    #[case(
+        &mut concat!(
+            "\n",
+            "% Note: This is a server message followed by an empty line\n"
+        )
+    )]
+    #[case(
+        &mut concat!(
+            "% Note: This is a server message preceding some newlines.\n",
+            "\n",
+            "\n",
+        )
+    )]
+    fn optional_comment_or_newlines_consumed(#[case] given: &mut &str) {
+        consume_opt_message_or_newlines(given).unwrap();
+        assert_eq!(*given, "");
+    }
+
+    #[test]
+    fn optional_comment_or_newlines_optional() {
+        assert_eq!(consume_opt_message_or_newlines(&mut ""), Ok(()));
+    }
 
     #[rstest]
     #[case(
